[RFC] create-namespace: Add percentage option
The size option of the create-namespace command can now take a percentage of
the total region size as the desired size. E.g. :

ndctl create-namespace  -s 50%

will create a namespace using half of the total region size, assuming
there is enough available capacity in the region for the request.

Only whole percentages are correctly parsed at the moment e.g. 12%, but
not 12.5%.

Sizes that result in a misaligned size with regards to region alignment
(default 16MB) are rounded down, unless the rounded-up size would result
in using the full remaining region capacity available.

Fixes: https://github.com/pmem/ndctl/issues/199
Signed-off-by: Vasilis Liaskovitis <vliaskovitis@suse.com>