ndctl: remove travis-ci
Drop the Travis CI integration. Their switch to '.com' and the
corresponding rework to billing plans and OSS 'credits' has introduced
unnecessary complexity to something that Just Worked before.

The only thing the CI was testing was builds on Ubuntu, but there
appears to be Debian (unstable) automation now to pick up and build new
upstream releases automatically [1], so we have some replacement coverage
by way of that.

[1]: https://packages.debian.org/source/sid/ndctl

Link: https://lore.kernel.org/r/20220830212338.413104-1-vishal.l.verma@intel.com
Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>