cxl/list: include --endpoints in -v listings
The first level of -v (--verbose) for cxl-list is meant to include all
topological details of the CXL hierarchy on a platform. The endpoints
listing was missing from the first level, and also wasn't added at a
higher verbosity level. Fix this omission by adding endpoints to the -v
verbosity level.

Fixes: 8b61e8e75443 ("cxl: Add list verbose option to the cxl command")
Cc: Matthew Ho <sunfishho12@gmail.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>