cxl/filter: enumerate endpoints and memdevs in an RCH
An RCH (Restricted CXL Host) is a platform limited to a smaller subset
of CXL functionality (formerly known as a CXL1.1 platform). Such a
playform may still be able to host a CXL device with a higher spec level
(section 9.11.8 CXL Devices Attached to an RCH in the CXL 3.0
specification).

The caveat with such a configuration is that the endpoints will appear
as 'Root Complex integrated Endpoints' (RCiEPs). In libcxl terms, the
endpoint would be connected directly to the cxl_bus object.

Until now, cxl-list and cxl_filter_walk() assume that endpoints must be
under a cxl_port, which is true for CXL VH topologies. To account for
endpoints in RCH topologies, change cxl_filter_walk() to also account
for endpoints that may appear directly under a cxl_bus object.

With this change, an example CXL RCH topology may enumerate as follows:

    # cxl list -v
    [
      {
        "bus":"root0",
        "provider":"ACPI.CXL",
        "nr_dports":1,
        "dports":[
          {
            "dport":"pci0000:38",
            "id":49
          }
        ],
        "endpoints:root0":[
          {
            "endpoint":"endpoint1",
            "host":"mem0",
            "depth":1,
            "memdev":{
              "memdev":"mem0",
              "pmem_size":0,
              "ram_size":17179869184,
              "serial":0,
              "numa_node":0,
              "host":"0000:38:00.0"
            },
            "decoders:endpoint1":[
              {
                "decoder":"decoder1.0",
                "resource":36507222016,
                "size":17179869184,
                "interleave_ways":1
              }
            ]
          }
        ]
      }
    ]

Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>