cxl: add a helper to go through all current events and parse them
Add a common function to iterate through and extract the events in the
current trace buffer. The function uses tracefs_iterate_raw_events() from
libtracefs to go through all the events loaded into a tep_handle. A
callback is provided to the API call in order to parse the event. For cxl
monitor, the "system" or category of trace event, in this case "cxl",
is provided in order to filter for the CXL events.

Tested-by: Alison Schofield <alison.schofield@intel.com>
Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/166803885992.145141.11751557821515668416.stgit@djiang5-desk3.ch.intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>