cxl/monitor: add a new monitor command for CXL trace events
Add function that creates an event trace instance and utilize the cxl event
trace common functions to extract matching events from the trace buffer.
The monitoring function will pend on an epoll fd and wait for new events
to appear in the trace buffer.

Connect the monitoring functionality to the cxl monitor command. Add basic
functionality to the cxl monitor command where it can be launched as a daemon
and logging can be designated to stdout or a file.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/166803887731.145141.12691742774548909895.stgit@djiang5-desk3.ch.intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>