util/trace: add an optional pid check to event parsing
When parsing events, callers may only be interested in events
that originate from the current process. Introduce an optional
argument to the event trace context: event_pid. When event_pid is
present, simply skip the parsing of events without a matching pid.
It is not a failure to see other, non matching events.

The initial use case for this is CXL device poison listings where
only the media-error records requested by this process are wanted.

Reviewed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/78e904ed934820f217f96d19603acf64e322184a.1720241079.git.alison.schofield@intel.com/
Signed-off-by: Alison Schofield <alison.schofield@intel.com>