cxl/test: add cxl_xor_region test
Exercise the kernel driver support of XOR math by creating regions
with 1,2, and 4-way interleaves using XOR interleave arithmetic.

Use module parameter "interleave_arithmetic=1" to select the cxl_test
topology that supports XOR math. XOR math is not used in the default
cxl_test module.

Add this test to the 'cxl' suite so that it gets exercised routinely.
If the topology defined in cxl_test changes, this test may require
an update.

Signed-off-by: Alison Schofield <alison.schofield@intel.com>
Link: https://lore.kernel.org/r/20221118045332.1330804-1-alison.schofield@intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>