meson.build: add a check argument to run_command
Meson has started to warn about:

  WARNING: You should add the boolean check kwarg to the run_command call.
         It currently defaults to false,
         but it will default to true in future releases of meson.
         See also: https://github.com/mesonbuild/meson/issues/9300

There is one instance of run_command() in the top-level meson.build
which elides the explicit check argument. Since we don't care about the
result of clean_config.sh (if any config.h are found they will be
cleaned, and if none are found, we're fine), add a 'check: false'
argument to this and squelch the warning.

Link: https://lore.kernel.org/r/20221206-vv-misc-v1-2-4c5bd58c90ca@intel.com
Cc: Dan Williams <dan.j.williams@intel.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>