cxl/list: Add a 'firmware_node' alias
In preparation for the kernel switching from ACPI0016 devices as the dport
identifier to the corresponding host-bridge as the identifier, add support
for listing the dport firmware_node as an alias, and determining when the
dev_path is the phys_path for a dport or a target.

The code paths that depend on phys_path like cxl_target_maps_memdev() and
cxl_dport_maps_memdev() already have appropriate fallbacks to dev_path.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Link: https://lore.kernel.org/r/167053491307.582963.8109215191948535895.stgit@dwillia2-xfh.jf.intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>