cxl/region: Fix memdev_filter_pos() separator detection
Stop using the modified version of the @_csv as the string to determine
the separator it may cause which_sep() to get different answers each
iteration. Moreover there is no need to retrieve the separator each
iteration, so just retrieve it once from @_csv.

If the separator is mismatched it can lead to the sort not following the
specified device order for create-region.

Reported-by: Luis Chamberlain <mcgrof@kernel.org>
Link: https://gist.github.com/mcgrof/aeac639365a651bd77f143cf38eb7493
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Link: https://lore.kernel.org/r/167106803771.1326377.10608704295797398750.stgit@dwillia2-xfh.jf.intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>