cxl/monitor: Make libtracefs dependency optional
Build a stub version of 'cxl monitor' that reports that the facility was
statically disabled at configure / build time. Provide the meson
configuration line to correct the build.

This is in response to the fact that some distros fail to ship
libtracefs-devel even though they ship libtracefs (looking at you CentOS
Stream 8).

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Link: https://lore.kernel.org/r/167718414197.16926.10394415062829848410.stgit@dwillia2-xfh.jf.intel.com
Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>