cxl/list: Fix filtering RCDs
Attempts to filter by memdev fail when the memdev is an RCD (RCH topology):

    # cxl list -BEM -m 11
      Warning: no matching devices found

    [
    ]

This is caused by VH topology assumption where an intervening host-bridge
port is expected between the root CXL port and the endpoint. In an RCH
topology an endpoint is integrated in the host-bridge.

Search for endpoints directly attached to the root:

    # cxl list -BEMu -m 11
    {
      "bus":"root3",
      "provider":"cxl_test",
      "endpoints:root3":[
        {
          "endpoint":"endpoint22",
          "host":"mem11",
          "depth":1,
          "memdev":{
            "memdev":"mem11",
            "ram_size":"2.00 GiB (2.15 GB)",
            "serial":"0xa",
            "numa_node":0,
            "host":"cxl_rcd.10"
          }
        }
      ]
    }

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Link: https://lore.kernel.org/r/168236637746.1027628.14674251843014155022.stgit@dwillia2-xfh.jf.intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>