ndctl: use strcmp for reserved word in monitor commands
According to the documentation, when '-l standard' is specified, the log
would be sent to stdout. But since it's using strncmp(a, b, 10) to
compare the first 10 characters, it will incorrectly detect a filename
starting with a substring 'standard' as stdout.

For example:

  $ cxl monitor -l standard.log

User is most likely want to save log to ./standard.log instead of stdout.

Change this to a 'strcmp' so that the whole string is considered.

Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Signed-off-by: Li Zhijian <lizhijian@fujitsu.com>
Link: https://lore.kernel.org/r/20230711115344.562823-4-lizhijian@fujitsu.com
Reviewed-by: Alison Schofield <alison.schofield@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>