ndctl/cxl/test: Add CXL event test
Previously CXL event testing was run by hand.  Improve testing coverage
and address a lack of automated regression testing by adding a unit test
for this.

Add a CXL event test as part of the meson test infrastructure.  Passing
is predicated on receiving the appropriate number of errors in each log.
Individual event values are not checked.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>
Link: https://lore.kernel.org/r/20230726-cxl-event-v2-1-550f5625d22f@intel.com
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>