cxl/region: Add -f option for disable-region
The current operation for disable-region does not check if the memory
covered by a region is online before attempting to disable the cxl region.
Have the tool attempt to offline the relevant memory before attempting to
disable the region(s). If offline fails, stop and return error.

Provide a -f option for the region to continue disable the region even if
the memory is not offlined. Add a warning to state that the physical
memory is being leaked and unrecoverable unless reboot due to disable without
offline.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/169878724592.82931.11180459815481606425.stgit@djiang5-mobl3
Reviewed-by: Xiao Yang <yangx.jy@fujitsu.com>
Reviewed-by: Alison Schofield <alison.schofield@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>