test/cxl-region-sysfs: fix a missing space syntax error
Currently the cxl-region-sysfs.sh test runs to completion and passes,
but with syntax errors in the log. It turns out that because the test is
checking for a positive condition as a failure, that also happens to
mask the syntax errors. Fix the syntax and note that this also happens
to unblock a test case that was being hidden by this error.

Signed-off-by: Li Zhijian <lizhijian@fujitsu.com>
Link: https://lore.kernel.org/r/20231213082556.1401741-2-lizhijian@fujitsu.com
Acked-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>