util/trace: add helpers to retrieve tep fields by type
Add helpers to extract the value of an event record field given the
field name. This is useful when the user knows the name and format
of the field and simply needs to get it. The helpers also return
the 'type'_MAX of the type when the field is

Since this is in preparation for adding a cxl_poison private parser
for 'cxl list --media-errors' support those specific required
types: u8, u32, u64.

Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Reviewed-by: Fan Ni <fan.ni@samsung.com>
Link: https://lore.kernel.org/r/b6089a98199539eca9c89f81de19cede18468408.1720241079.git.alison.schofield@intel.com/
Signed-off-by: Alison Schofield <alison.schofield@intel.com>